+++ packages/vexide-devices/src/lib.rs
@@ -22,10 +22,10 @@ pub mod smart;
 pub mod battery;
 pub mod color;
 pub mod controller;
+pub mod display;
 pub mod geometry;
 pub mod peripherals;
 pub mod position;
-pub mod screen;
 pub mod usd;
 
 use snafu::Snafu;